refactor: Post 내부 구조 SummaryCard 컴포넌트로 분리 [feat/21]
