+++ src/components/card/index.js
@@ -1,5 +1,6 @@
 export * from "./cardFooter";
 export * from "./contents";
+export { default as SummaryCard } from "./SummaryCard";
 export { default as FullCard } from "./FullCard";
 export { default as Card } from "./Card";
 export { default as Content } from "./Content";
